Redo shop_watcher module to be asynchronous
Merged the previous client functionality directly into the scanner module, and made the functions async. I've been told it would work better with websockets and while I can see why it would be faster, I don't really get it beyond that, for now.
